Add methods to store recently viewed tickets by user
In order to allow recently viewed tickets to be stored for each user, two new methods are added.

The method RT::User->RecentlyViewedTickets returns an ordered list of RT::Ticket objects.
The order of the list is determined by the second method.

The second method RT::User->AddRecentlyViewedTicket takes a single RT::Ticket object and adds the ticket to a stored list based on the current time. The method also truncates the list of tickets to be no greater then ten and handles moving existing tickets to the beginning of the list.