Add shared method to update recently viewed tickets
In order to simplify the different ways to view a ticket, a new mason template is added to update a users recently viewed tickets.

In order to update the users recently viewed tickets the BecomeRecent template is loaded in four major pages where a ticket is displayed with the intention of being viewed. These pages include Display.html, History.html, Modify.html and ModifyAll.html.

Fixes: I#18820